car-management/docker-dev-lpr/src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
flask==1.0.2
flask-cors==3.0.4
h5py==2.7.1
imageio==2.3.0
keras==2.1.5
pillow==4.0.0
numpy==1.13.3
tensorflow==1.7.1
scipy==1.0.1
pandas==0.22.0
xlrd
